lttb: Fix LTTB bucketing and add better tests
